+++ CHANGELOG.md
@@ -49,6 +49,7 @@ and this project adheres to
 
 ### Changed
 
+- Client objects in the configuration file are now sorted ([#3933]).
 - Responses from cache are now labeled ([#3772]).
 - Better error message for ED25519 private keys, which are not widely supported
   ([#3737]).
@@ -232,6 +233,7 @@ In this release, the schema version has changed from 10 to 12.
 [#3772]: https://github.com/AdguardTeam/AdGuardHome/issues/3772
 [#3815]: https://github.com/AdguardTeam/AdGuardHome/issues/3815
 [#3890]: https://github.com/AdguardTeam/AdGuardHome/issues/3890
+[#3933]: https://github.com/AdguardTeam/AdGuardHome/pull/3933